moved balance test to separate file
